ABAC-Tutorial
Here's what the script does:

Defines a list of users to be created.
Loops through each user in the list and executes the following steps:
Creates the IAM user using the aws iam create-user command.
Attaches the AWSAccessKeys policy to the user using the aws iam attach-user-policy command. This policy grants the user permission to create, modify, and delete their own access keys, which are necessary for the user to assume roles later on.
Tags the user with the appropriate tags using the aws iam tag-user command. The script extracts the values for the tags from the user name by using the cut command to split the string at the hyphens.
Please note that you need to replace aws:iam::aws:policy/AWSAccessKeys with the appropriate policy ARN for your environment if you are using a custom policy for managing access keys. Also, make sure to replace us-east-1 with the appropriate region if you want to create the users in a different region.